+++ src/components/fallbacks/display-trip-fallback.tsx
