+++ src/components/trips/display-trip/DisplayTrip.tsx
