+++ src/components/trips/display-trip/displayTripTable.tsx
