+++ src/components/trips/display-trips/trips-table.tsx
